Update README.md
I added the HubL striptags filter to the initialPostWords variable. The reason was that if a post has embedded content, this snippet of code, sans-striptags, was causing the reading minutes to be extremely high. It was calculating all of that HTML as words as well. Stripping out this HTML before we calculate the actual reading minutes is needed in order to get a better and more accurate number.